+++ flytekit/extras/external_container_sqlite3/task.py
@@ -11,7 +11,8 @@ import pandas as pd
 from flytekit import FlyteContext, kwtypes
 from flytekit.core.base_sql_task import SQLTask
 from flytekit.core.context_manager import SerializationSettings
-from flytekit.core.python_third_party_task import PythonThirdPartyContainerTask, ShimTaskExecutor
+from flytekit.core.python_third_party_task import PythonThirdPartyContainerTask
+from flytekit.core.shim_task import ShimTaskExecutor
 from flytekit.models import task as task_models
 from flytekit.types.schema import FlyteSchema
