+++ tests/flytekit/unit/extras/external_container_sqlite3/test_task.py
@@ -91,7 +91,7 @@ def test_task_serialization():
         "flytekit.core.python_third_party_task.default_task_template_resolver",
         "--",
         "{{.taskTemplatePath}}",
-        "flytekit.extras.sqlite3.task.SQLite3TaskExecutor",
+        "flytekit.extras.external_container_sqlite3.task.SQLite3TaskExecutor",
     ]
 
     assert tt.custom["query_template"] == "select TrackId, Name from tracks limit {{.inputs.limit}}"
